The Start of a Long Journey (Generator "Done")
The Generator functionality is, for all intents and purposes, done. There are some security flaws that definitely need to be improved, there are some other concepts in use that can be improved, and there is some extra error handling I would like to implement. However, the core functionality of the Generate function is there.