Fixed an Issue, Made Improvements, Fixed Spelling
Removed comments, solved an issue where using a specific special character set and then generating a random and then familiar password right after would cause an issue, due to a stupid mistake of looking for the wrong string (it was always wrong). Fixed some spelling and added another question or two to the questions to encourage/provoke thought.